Extract parameters to rocket routes into a common type
we move the most common (5) parameters into a new type, `impl`
FromRequest for it, and start using it to reduce the amount of
parameters in our route functions.

Unusual for this patch: I use an early return, to move the main logic
out of giant indented `else` into the body of the function.

n.b.: even tho clippy did *not* complain about this particular route
function, we still use PlumeRockets for consistency.